Implemented mapping ActiveResource::Base classes. The implementation is not DRY, but preserves the current speed of ActiveRecord serialization as the trade-off.
Since ActiveResource#save creates a new record or updates an existing record, rubyamf save functionality is maintained automatically. Association mapping for active_resource is not implemented, but I believe they could be implemented in the associated controller of the ActiveResource model so that they pass in an out as attributes of the ActiveResource instance.